R Code from SCCWRP with Comments
I only added the comments to one strata. It gets redundant as the same process is applied. There are a few things to call out as well as a question. Search karen to locate qs. 
TOC almost always defined with incorrect parameter. I changed this to reflect the correct parameter. Also happened once with zinc and copper. 

Did not go through the plots at the end.